Change SectorPresenter to support a 'start' arg
This changes the SectorPresenter to support a new argument providing
the offset/'start' value for the list of recently changed documents.

The 'start' value is then passed as an argument to `LatestChanges.find`
where, as part of the behaviour added in the previous commit, it is 
included in the search query to Rummager.

This includes a new test helper for the integration tests, which creates
a stub for Rummager only invoked when the correct 'start' parameter is 
included.