Include pagination metadata in API response
This changes the SectorPresenter so that the values of the 'start' 
and 'total' keys in the latest changes response from Rummager are 
exposed in the API response.

We will need to access these values in order to correctly link to the
next and previous result pages, and to determine whether are additional 
pages or not.