Support the 'count' parameter
Allow a 'count' to be specified when requesting the list of latest
changes for a sub-topic. 

Like 'start', the value of 'count' is used to build the request to 
Rummager. 

We're restricting the value to be greater than zero and less
than or equal to 100, which matches the restrictions in Rummager. We 
continue to default to a count of 50 if the parameter is missing, blank, 
negative or zero.