Fix display of topic with empty groups in content-store.
Now that we're sending items to the content-store for all topics, the
ones with no curated groups have an empty groups array.  The log here
wasn't handling that case, and therefore assumed that they had curated
groups, and returned an empty list of groups.  This resulted in a blank
page on the frontend.